requirements.txt
pywin32
opencv_python
pyyaml
requests